Changed the design of messages and mechanisms for working with it. Added a multifunctional facade basic_actor.

+++ libactor_zeta_core/actor-zeta/messaging/message_priority.hpp
@@ -6,7 +6,7 @@ namespace actor_zeta {
 ///
 /// @enum
 ///
-        enum class message_priority : int {
+        enum class message_priority : uint8_t {
             low = 0,
             normal,
             high
